+++ src/data/Player.java
@@ -193,7 +193,7 @@ public class Player
 	 * @param effect the given effect
 	 * @param seconds the duration of the effect
 	 */
-	public void addStatus(Status effect, long seconds)
+	public void addStatus(Status effect, float seconds)
 	{
 		statuses.addStatus(effect, seconds);
 	}
